[MAJOR] Make it possible to use Vid on Vimeo
On Vimeo, the main video object is wrapped in some blocks, but they are
of the same dimensions, so user can't select them with cursor.
Moreover, a sibling of the topmost of these blocks is a block containing
another object (the scrollbar near "Other videos"), so it is impossible
to select a single object.

To deal with this, a way to iterate object on the page is provided.
User is now able to select a block with cursor and then use Space or Tab
to iterate objects lying underneath it.  For convenience, Enter can now
be used for confirming selection (as addition to left click).